some refactor

+++ server/database/migrations/1568864068834_subsidiary_schema.js
@@ -11,7 +11,7 @@ class SubsidiarySchema extends Schema {
       table.string('state', 100).notNullable()
       table.string('city', 100).notNullable()
       table.string('parish', 100).notNullable()
-      table.integer('distanceFromCaracas').notNullable()
+      table.float('distanceFromCaracas').notNullable()
     })
   }
